+++ tests/providers/azure/services/app/app_function_access_keys_configured/app_function_access_keys_configured_test.py
@@ -87,6 +87,7 @@ class Test_app_function_access_keys_configured:
                         public_access=False,
                         vnet_subnet_id=None,
                         ftps_state="AllAllowed",
+                        resource_group_name="resource_group_name",
                     )
                 }
             }
@@ -142,6 +143,7 @@ class Test_app_function_access_keys_configured:
                         public_access=False,
                         vnet_subnet_id=None,
                         ftps_state="AllAllowed",
+                        resource_group_name="resource_group_name",
                     )
                 }
             }