+++ tests/providers/azure/services/app/app_function_application_insights_enabled/app_function_application_insights_enabled_test.py
@@ -95,6 +95,7 @@ class Test_app_function_application_insights_enabled:
                         public_access=False,
                         vnet_subnet_id=None,
                         ftps_state="AllAllowed",
+                        resource_group_name="resource_group_name",
                     )
                 }
             }
@@ -164,6 +165,7 @@ class Test_app_function_application_insights_enabled:
                         public_access=False,
                         vnet_subnet_id=None,
                         ftps_state="AllAllowed",
+                        resource_group_name="resource_group_name",
                     )
                 }
             }
@@ -233,6 +235,7 @@ class Test_app_function_application_insights_enabled:
                         public_access=False,
                         vnet_subnet_id=None,
                         ftps_state="AllAllowed",
+                        resource_group_name="resource_group_name",
                     )
                 }
             }
@@ -302,6 +305,7 @@ class Test_app_function_application_insights_enabled:
                         public_access=False,
                         vnet_subnet_id=None,
                         ftps_state="AllAllowed",
+                        resource_group_name="resource_group_name",
                     )
                 }
             }